requirements.txt
requests
python-dotenv
pandas
seaborn
sklearn
matplotlib
statsmodels
bs4
re